fixed problems with cython on the following architectures:
- Linux laptok 2.6.31-ARCH #1 SMP PREEMPT Tue Nov 10 19:48:17 CET 2009 i686
  Intel(R) Core(TM) Duo CPU T2350 @ 1.86GHz GenuineIntel GNU/Linux

- Linux struc5 2.6.31-ARCH #1 SMP PREEMPT Tue Nov 10 19:01:40 CET 2009 x86_64
  Intel(R) Core(TM)2 Extreme CPU X9650 @ 3.00GHz GenuineIntel GNU/Linux

Using:

- Python2.6
- Cython 0.11.3

I kindly ask with different set-ups (Mac OSX, FreeBSD, Solaris etc.) to test
this. And run the following tests:

$ cd pycogent/tests; python test_maths/test_spatial/test_ckd3.py
....
----------------------------------------------------------------------
Ran 4 tests in 0.008s

OK
$ cd pycogent/tests; python test_struct/test_contact.py
.....
----------------------------------------------------------------------
Ran 5 tests in 1.256s

OK
$ cd pycogent/tests; python test_struct/test_asa.py
..........
----------------------------------------------------------------------
Ran 10 tests in 8.717s

OK

+++ cogent/struct/contact.py
@@ -7,7 +7,7 @@ from cogent.maths.geometry import coords_to_symmetry, \
                                   coords_to_crystal
 from _contact import cnt_loop
 from collections import defaultdict
-from numpy import array, r_, sqrt
+from numpy import array, r_, sqrt, int64
 
 
 __author__ = "Marcin Cieslik"
@@ -126,8 +126,8 @@ def _prepare_contacts(query, model=None, level='A', search_limit=6.0, \
         lc.append(chain2id[lent_id[2]])
     for qent_id in qents_ids:
         qc.append(chain2id[qent_id[2]])
-    lc = array(lc)
-    qc = array(qc)
+    lc = array(lc, dtype=int64)
+    qc = array(qc, dtype=int64)
     # here we leave python
     (idxc, n_src, n_asu, n_sym, n_tra, n_dst) = cnt_loop(\
                             qcoords, lcoords, qc, lc, shape[1], shape[2], \
